platform/packages/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7",
    "fastapi>=0.111",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
